+++ frontend/src/components/index.js
@@ -6,6 +6,7 @@ import Form from './Form.svelte';
 import Input from './Input.svelte';
 import ItemDisplay from './ItemDisplay.svelte';
 import Loading from './Loading.svelte';
+import NewArrivals from './NewArrivals.svelte';
 import RoleSelect from './RoleSelect.svelte';
 import StudentSearch from './StudentSearch.svelte';
 import Toast from './Toast.svelte';
@@ -21,6 +22,7 @@ export {
 	Input,
 	ItemDisplay,
 	Loading,
+	NewArrivals,
 	RoleSelect,
 	StudentSearch,
 	Toast,
